+++ Runtime/Utility/WebEventSystem.cs
@@ -1,10 +1,13 @@
 using UnityEngine.EventSystems;
 
-namespace YandexGames
+namespace YandexGames.Utility
 {
     /// <summary>
-    /// Workaround for EventSystem bug https://trello.com/c/PjW4j3st
+    /// Fixes unresponsive UI controls after alt-tabbing on mobile Google Chrome.
     /// </summary>
+    /// <remarks>
+    /// Workaround for <see href="https://trello.com/c/PjW4j3st"/>
+    /// </remarks>
     public class WebEventSystem : EventSystem
     {
         protected override void OnApplicationFocus(bool hasFocus) => base.OnApplicationFocus(true);