+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v6.0.0  
+Breaking changes:  
+- PlayerAccount.Authorized is now PlayerAccount.IsAuthorized. Needed that to achieve full naming consistency across internal and external API.  
+- Same goes for YandexGamesSdk.IsInitialized.  
+  
 v5.0.1  
 Fixes:  
 - Fixed Leaderboard.GetPlayerEntry successCallback silently stopping execution when it's supposed to return null.